+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/Android.kt
@@ -26,7 +26,6 @@ import com.getstream.sdk.chat.viewmodel.messages.MessageListViewModel
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.QueryChannelRequest
 import io.getstream.chat.android.client.api.models.QuerySort
-import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Channel
@@ -35,7 +34,6 @@ import io.getstream.chat.android.client.models.Filters
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.livedata.ChatDomain
-import io.getstream.chat.android.livedata.utils.RetryPolicy
 import io.getstream.chat.android.ui.ChatUI
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.StyleTransformer
@@ -675,20 +673,21 @@ class Android {
             val chatDomain = ChatDomain.instance()
         }
 
-        fun initializeChatDomainWithCustomRetryPolicy() {
-            val chatClient = ChatClient.Builder("apiKey", requireContext()).build()
-            val chatDomain = ChatDomain.Builder(requireContext(), chatClient)
-                .retryPolicy(object : RetryPolicy {
-                    override fun shouldRetry(client: ChatClient, attempt: Int, error: ChatError): Boolean {
-                        return attempt < 3
-                    }
-
-                    override fun retryTimeout(client: ChatClient, attempt: Int, error: ChatError): Int {
-                        return 1000 * attempt
-                    }
-                })
-                .build()
-        }
+        // TODO: ChatDomain docs will be removed in scope of https://github.com/GetStream/stream-chat-android/issues/3034
+        // fun initializeChatDomainWithCustomRetryPolicy() {
+        //     val chatClient = ChatClient.Builder("apiKey", requireContext()).build()
+        //     val chatDomain = ChatDomain.Builder(requireContext(), chatClient)
+        //         .retryPolicy(object : RetryPolicy {
+        //             override fun shouldRetry(client: ChatClient, attempt: Int, error: ChatError): Boolean {
+        //                 return attempt < 3
+        //             }
+        //
+        //             override fun retryTimeout(client: ChatClient, attempt: Int, error: ChatError): Int {
+        //                 return 1000 * attempt
+        //             }
+        //         })
+        //         .build()
+        // }
 
         fun watchChannel() {
             val chatDomain = ChatDomain.instance()