+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/ChatDomainImpl.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.os.Handler
 import androidx.annotation.CheckResult
 import androidx.annotation.VisibleForTesting
+import io.getstream.chat.android.client.BuildConfig
 import io.getstream.chat.android.client.BuildConfig.STREAM_CHAT_VERSION
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.FilterObject
@@ -19,15 +20,14 @@ import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.events.MarkAllReadEvent
 import io.getstream.chat.android.client.extensions.cidToTypeAndId
 import io.getstream.chat.android.client.extensions.enrichWithCid
+import io.getstream.chat.android.client.extensions.isPermanent
 import io.getstream.chat.android.client.logger.ChatLogger
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Channel
-import io.getstream.chat.android.client.models.ChannelMute
 import io.getstream.chat.android.client.models.Config
 import io.getstream.chat.android.client.models.Filters.`in`
 import io.getstream.chat.android.client.models.Member
 import io.getstream.chat.android.client.models.Message
-import io.getstream.chat.android.client.models.Mute
 import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.TypingEvent
 import io.getstream.chat.android.client.models.User
@@ -38,20 +38,18 @@ import io.getstream.chat.android.client.utils.map
 import io.getstream.chat.android.client.utils.observable.Disposable
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
-import io.getstream.chat.android.livedata.BuildConfig
 import io.getstream.chat.android.offline.channel.ChannelController
 import io.getstream.chat.android.offline.event.EventHandlerImpl
 import io.getstream.chat.android.offline.experimental.channel.state.toMutableState
 import io.getstream.chat.android.offline.experimental.channel.thread.state.toMutableState
 import io.getstream.chat.android.offline.experimental.global.GlobalMutableState
-import io.getstream.chat.android.offline.experimental.global.toMutableState
-import io.getstream.chat.android.offline.experimental.plugin.OfflinePlugin
+import io.getstream.chat.android.offline.experimental.global.GlobalState
+import io.getstream.chat.android.offline.experimental.plugin.logic.LogicRegistry
+import io.getstream.chat.android.offline.experimental.plugin.state.StateRegistry
 import io.getstream.chat.android.offline.experimental.querychannels.state.toMutableState
 import io.getstream.chat.android.offline.extensions.applyPagination
 import io.getstream.chat.android.offline.extensions.cancelMessage
 import io.getstream.chat.android.offline.extensions.createChannel
-import io.getstream.chat.android.offline.extensions.isPermanent
-import io.getstream.chat.android.offline.extensions.loadMessageById
 import io.getstream.chat.android.offline.extensions.loadOlderMessages
 import io.getstream.chat.android.offline.extensions.sendGiphy
 import io.getstream.chat.android.offline.extensions.shuffleGiphy
@@ -87,10 +85,7 @@ import io.getstream.chat.android.offline.usecase.SendMessage
 import io.getstream.chat.android.offline.usecase.SendReaction
 import io.getstream.chat.android.offline.usecase.ShowChannel
 import io.getstream.chat.android.offline.usecase.WatchChannel
-import io.getstream.chat.android.offline.utils.CallRetryService
-import io.getstream.chat.android.offline.utils.DefaultRetryPolicy
 import io.getstream.chat.android.offline.utils.Event
-import io.getstream.chat.android.offline.utils.RetryPolicy
 import io.getstream.chat.android.offline.utils.validateCid
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
@@ -143,10 +138,9 @@ internal class ChatDomainImpl internal constructor(
     override var userPresence: Boolean = false,
     internal var backgroundSyncEnabled: Boolean = false,
     internal var appContext: Context,
-    private val offlinePlugin: OfflinePlugin,
     internal val uploadAttachmentsNetworkType: UploadAttachmentsNetworkType = UploadAttachmentsNetworkType.NOT_ROAMING,
-    override val retryPolicy: RetryPolicy = DefaultRetryPolicy(),
-) : ChatDomain {
+    private val globalState: GlobalMutableState = GlobalMutableState.getOrCreate()
+) : ChatDomain, GlobalState by globalState {
     internal constructor(
         client: ChatClient,
         handler: Handler,
@@ -155,7 +149,7 @@ internal class ChatDomainImpl internal constructor(
         userPresence: Boolean,
         backgroundSyncEnabled: Boolean,
         appContext: Context,
-        offlinePlugin: OfflinePlugin,
+        globalState: GlobalMutableState = GlobalMutableState.getOrCreate()
     ) : this(
         client = client,
         db = null,
@@ -165,8 +159,11 @@ internal class ChatDomainImpl internal constructor(
         userPresence = userPresence,
         backgroundSyncEnabled = backgroundSyncEnabled,
         appContext = appContext,
-        offlinePlugin = offlinePlugin,
+        globalState = globalState
     )
+
+    private val state: StateRegistry by lazy { StateRegistry.getOrCreate(scope, user, repos, latestUsers) }
+    private val logic: LogicRegistry by lazy { LogicRegistry.getOrCreate(state) }
 
     // Synchronizing ::retryFailedEntities execution since it is called from multiple places. The shared resource is DB.stream_chat_message table.
     private val entitiesRetryMutex = Mutex()
@@ -192,49 +189,6 @@ internal class ChatDomainImpl internal constructor(
         }
 
     private var _repos: RepositoryFacade? = null
-
-    private val globalState: GlobalMutableState = offlinePlugin.globalState.toMutableState()
-
-    override val user: StateFlow<User?> by globalState::user
-
-    /** if the client connection has been initialized */
-    override val initialized: StateFlow<Boolean> = globalState.initialized
-
-    /**
-     * StateFlow<Boolean> that indicates if we are currently online
-     */
-    override val connectionState: StateFlow<ConnectionState> = globalState.connectionState
-
-    /**
-     * The total unread message count for the current user.
-     * Depending on your app you'll want to show this or the channelUnreadCount
-     */
-    override val totalUnreadCount: StateFlow<Int> = globalState.totalUnreadCount
-
-    /**
-     * the number of unread channels for the current user
-     */
-    override val channelUnreadCount: StateFlow<Int> = globalState.channelUnreadCount
-
-    /**
-     * list of users that you've muted
-     */
-    override val muted: StateFlow<List<Mute>> = globalState.muted
-
-    /**
-     * List of channels you've muted
-     */
-    override val channelMutes: StateFlow<List<ChannelMute>> = globalState.channelMutes
-
-    /**
-     * if the current user is banned or not
-     */
-    override val banned: StateFlow<Boolean> = globalState.banned
-
-    /**
-     * The error event state flow object is triggered when errors in the underlying components occur.
-     */
-    override val errorEvents: StateFlow<Event<ChatError>> = globalState.errorEvents
 
     /** the event subscription */
     private var eventSubscription: Disposable = EMPTY_DISPOSABLE
@@ -387,7 +341,8 @@ internal class ChatDomainImpl internal constructor(
         eventHandler.clear()
         activeChannelMapImpl.clear()
         activeQueryMapImpl.clear()
-        offlinePlugin.clear()
+        logic.clear()
+        state.clear()
     }
 
     override fun getVersion(): String {
@@ -401,14 +356,6 @@ internal class ChatDomainImpl internal constructor(
     private fun initClean() {
         mainHandler.postDelayed(cleanTask, 5000)
     }
-
-    internal fun callRetryService() = CallRetryService(retryPolicy, client)
-
-    @Deprecated(
-        message = "This utility method is extracted to CallRetryService",
-        replaceWith = ReplaceWith("ChatDomainImpl::callRetryService::runAndRetry")
-    )
-    suspend fun <T : Any> runAndRetry(runnable: () -> Call<T>): Result<T> = callRetryService().runAndRetry(runnable)
 
     fun addError(error: ChatError) {
         globalState._errorEvent.value = Event(error)
@@ -480,8 +427,8 @@ internal class ChatDomainImpl internal constructor(
         val cid = "%s:%s".format(channelType, channelId)
         if (!activeChannelMapImpl.containsKey(cid)) {
             val channelController = ChannelController(
-                mutableState = offlinePlugin.state.channel(channelType, channelId).toMutableState(),
-                channelLogic = offlinePlugin.logic.channel(channelType, channelId),
+                mutableState = state.channel(channelType, channelId).toMutableState(),
+                channelLogic = logic.channel(channelType, channelId),
                 client = client,
                 domainImpl = this,
             )
@@ -540,8 +487,8 @@ internal class ChatDomainImpl internal constructor(
         sort: QuerySort<Channel>,
     ): QueryChannelsController =
         activeQueryMapImpl.getOrPut("${filter.hashCode()}-${sort.hashCode()}") {
-            val mutableState = offlinePlugin.state.queryChannels(filter, sort).toMutableState()
-            val logic = offlinePlugin.logic.queryChannels(filter, sort)
+            val mutableState = state.queryChannels(filter, sort).toMutableState()
+            val logic = logic.queryChannels(filter, sort)
             QueryChannelsController(domainImpl = this, mutableState = mutableState, queryChannelsLogic = logic)
         }
 
@@ -893,8 +840,8 @@ internal class ChatDomainImpl internal constructor(
         return CoroutineCall(scope) {
             Result.success(
                 channel(cid).getThread(
-                    offlinePlugin.state.thread(parentId).toMutableState(),
-                    offlinePlugin.logic.thread(parentId)
+                    state.thread(parentId).toMutableState(),
+                    logic.thread(parentId)
                 )
             )
         }