+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/global/GlobalMutableState.kt
@@ -2,6 +2,7 @@
 
 package io.getstream.chat.android.offline.experimental.global
 
+import androidx.annotation.VisibleForTesting
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.models.ChannelMute
 import io.getstream.chat.android.client.models.Mute
@@ -13,7 +14,7 @@ import io.getstream.chat.android.offline.utils.Event
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
-internal class GlobalMutableState : GlobalState {
+internal class GlobalMutableState private constructor() : GlobalState {
 
     internal val _initialized = MutableStateFlow(false)
     internal val _connectionState = MutableStateFlow(ConnectionState.OFFLINE)
@@ -56,6 +57,26 @@ internal class GlobalMutableState : GlobalState {
 
     override fun isInitialized(): Boolean {
         return _initialized.value
+    }
+
+    internal companion object {
+        private var instance: GlobalMutableState? = null
+
+        /**
+         * Gets the singleton of [GlobalMutableState] or creates it in the first call.
+         */
+        internal fun getOrCreate(): GlobalMutableState {
+            return instance ?: GlobalMutableState().also { globalState ->
+                instance = globalState
+            }
+        }
+
+        /**
+         * Creates an instance of [GlobalMutableState] with a fresh state. Please keep in mind that many instances of this class may
+         * cause the SDK to present an inconsistent state.
+         */
+        @VisibleForTesting
+        internal fun create(): GlobalMutableState = GlobalMutableState()
     }
 }